requirements.txt
fastapi
uvicorn[standard]
python-dotenv
langchain-ollama
langchain-tavily
langgraph
pydantic
psycopg[binary]
langgraph-checkpoint-postgres